Introduce base entity in Open Router
