+++ homeassistant/components/open_router/conversation.py
@@ -2,24 +2,14 @@
 
 from typing import Literal
 
-import openai
-from openai.types.chat import (
-    ChatCompletionAssistantMessageParam,
-    ChatCompletionMessageParam,
-    ChatCompletionSystemMessageParam,
-    ChatCompletionUserMessageParam,
-)
-
 from homeassistant.components import conversation
-from homeassistant.config_entries import ConfigSubentry
-from homeassistant.const import CONF_LLM_HASS_API, CONF_MODEL, MATCH_ALL
+from homeassistant.const import CONF_LLM_HASS_API, MATCH_ALL
 from homeassistant.core import HomeAssistant
-from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import OpenRouterConfigEntry
-from .const import DOMAIN, LOGGER
+from .const import DOMAIN
+from .entity import OpenRouterEntity
 
 
 async def async_setup_entry(
@@ -35,46 +25,10 @@ async def async_setup_entry(
         )
 
 
-def _convert_content_to_chat_message(
-    content: conversation.Content,
-) -> ChatCompletionMessageParam | None:
-    """Convert any native chat message for this agent to the native format."""
-    LOGGER.debug("_convert_content_to_chat_message=%s", content)
-    if isinstance(content, conversation.ToolResultContent):
-        return None
-
-    role: Literal["user", "assistant", "system"] = content.role
-    if role == "system" and content.content:
-        return ChatCompletionSystemMessageParam(role="system", content=content.content)
-
-    if role == "user" and content.content:
-        return ChatCompletionUserMessageParam(role="user", content=content.content)
-
-    if role == "assistant":
-        return ChatCompletionAssistantMessageParam(
-            role="assistant", content=content.content
-        )
-    LOGGER.warning("Could not convert message to Completions API: %s", content)
-    return None
-
-
-class OpenRouterConversationEntity(conversation.ConversationEntity):
+class OpenRouterConversationEntity(OpenRouterEntity, conversation.ConversationEntity):
     """OpenRouter conversation agent."""
 
-    _attr_has_entity_name = True
     _attr_name = None
-
-    def __init__(self, entry: OpenRouterConfigEntry, subentry: ConfigSubentry) -> None:
-        """Initialize the agent."""
-        self.entry = entry
-        self.subentry = subentry
-        self.model = subentry.data[CONF_MODEL]
-        self._attr_unique_id = subentry.subentry_id
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, subentry.subentry_id)},
-            name=subentry.title,
-            entry_type=DeviceEntryType.SERVICE,
-        )
 
     @property
     def supported_languages(self) -> list[str] | Literal["*"]:
@@ -86,7 +40,7 @@ class OpenRouterConversationEntity(conversation.ConversationEntity):
         user_input: conversation.ConversationInput,
         chat_log: conversation.ChatLog,
     ) -> conversation.ConversationResult:
-        """Process a sentence."""
+        """Process the user input and call the API."""
         options = self.subentry.data
 
         try:
@@ -99,35 +53,6 @@ class OpenRouterConversationEntity(conversation.ConversationEntity):
         except conversation.ConverseError as err:
             return err.as_conversation_result()
 
-        messages = [
-            m
-            for content in chat_log.content
-            if (m := _convert_content_to_chat_message(content))
-        ]
-
-        client = self.entry.runtime_data
-
-        try:
-            result = await client.chat.completions.create(
-                model=self.model,
-                messages=messages,
-                user=chat_log.conversation_id,
-                extra_headers={
-                    "X-Title": "Home Assistant",
-                    "HTTP-Referer": "https://www.home-assistant.io/integrations/open_router",
-                },
-            )
-        except openai.OpenAIError as err:
-            LOGGER.error("Error talking to API: %s", err)
-            raise HomeAssistantError("Error talking to API") from err
-
-        result_message = result.choices[0].message
-
-        chat_log.async_add_assistant_content_without_tools(
-            conversation.AssistantContent(
-                agent_id=user_input.agent_id,
-                content=result_message.content,
-            )
-        )
+        await self._async_handle_chat_log(user_input, chat_log)
 
         return conversation.async_get_result_from_chat_log(user_input, chat_log)
